Moved `prisonerLocation()` tests
Into `prisonerLocationUtils.test.ts` which is where the
tests for the `utils/prisonerLocationUtils` module should
live.
